requirements.txt
google-genai>=1.14.0
apify-client>=1.9.0
elevenlabs>=1.50.0
audioop-lts>=0.2.1
pydub>=0.25.1
python-dotenv>=1.1.0
flask>=3.1.0
